==> R/05_make_spreadsheet.R <==
# Write all tables to website/all_tables.xlsx

filename <- "website/tables2022.xlsx"
period <- "2018-2020"

# Prelims ----------------------------------------------------------------------
library(tidyverse)
library(openxlsx)

source("R/00_functions.R", encoding = "UTF-8")

# load tables
tables1 <- readRDS("data/tables1.rds")
tables2 <- readRDS("data/tables2.rds")
tables3 <- readRDS("data/tables3.rds")
tables4 <- readRDS("data/tables4.rds")
tables5 <- readRDS("data/tables5.rds")

all_tables <- c(tables1, tables2, tables3, tables4, tables5)

tables_n <- length(all_tables)

# Create all headers -----------------------------------------------------------

# each list item in headers is for one worksheet / one chart in report
# headers$header contains worksheet title, any notes, and data source
# headers$titles contains table titles for all tables in a worksheet

headers <- list(tables_n)

# * 1 Total wealth -------------------------------------------------------------

# tables 1-4

headers <- list(list(header = c("Median household wealth",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Note: 2006-2008 estimates are based on the half of the sample that were asked the physical wealth questions.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) household wealth in £, Scotland",
                                "Sample sizes, Scotland")),
                list(header = c("Household wealth by type of wealth",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Note: Includes all households, whether or not they have each type of wealth.",
                                "Source: Wealth and Assets Survey"),
                     titles = c(paste0("Median household wealth in £ by type of wealth, Scotland ", period),
                                paste0("Sample sizes by type of wealth, Scotland ", period))),
                list(header = c("Household wealth by type of wealth (time series)",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Note: Includes all households, whether or not they have each type of wealth.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) household wealth in £ by type of wealth, Scotland",
                                "Sample sizes by type of wealth, Scotland")),
                list(header = c("Household wealth by wealth type and age",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c(paste0("Median household wealth in £ by type of wealth and age of household head, Scotland ", period),
                                paste0("Sample sizes by type of wealth and age of household head, Scotland ", period))),
                
# * 2 Wealth inequality --------------------------------------------------------
                
# tables 5-26
                
                list(header = c("Wealth inequality and income inequality",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c(paste0("Share of total wealth (negative wealth set to zero) and of total household income in each 2% wealth / income band, Scotland ", period),
                                paste0("Sample sizes, Scotland ", period))),
                list(header = c("Household wealth by wealth decile",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c(paste0("Median household wealth in £ by wealth decile, Scotland ", period), 
                                paste0("Median age of household head by wealth decile, Scotland ", period), 
                                paste0("Sample sizes by wealth decile, Scotland ", period))),
                list(header = c("Household wealth by wealth decile (time series)",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: 2006-2008 estimates are based on the half of the sample that were asked the physical wealth questions.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) household wealth in £ by wealth decile, Scotland",
                                "Median age of household head by wealth decile, Scotland",
                                "Sample sizes by wealth decile, Scotland")),
                list(header = c("Age groups in each wealth decile",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c(paste0("Share of age groups (age of household head) in each wealth decile, Scotland ", period),
                                paste0("Sample sizes, Scotland ", period))),
                list(header = c("Gini coefficient of wealth inequality",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Note: 2006-2008 estimates are based on the half of the sample that were asked the physical wealth questions.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Gini coefficient of wealth inequality, Scotland",
                                "Sample sizes, Scotland")),
                list(header = c("Palma ratio of wealth inequality",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Note: 2006-2008 estimates are based on the half of the sample that were asked the physical wealth questions.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Palma ratio of wealth inequality, Scotland",
                                "Sample sizes, Scotland")),
                list(header = c("Total wealth of the 40% least wealthy, the middle 50%, the 10% wealthiest and the 1% wealthiest households",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Note: 2006-2008 estimates are based on the half of the sample that were asked the physical wealth questions.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Share of total wealth of the 40% least wealthy, the middle 50%, the 10% wealthiest and the 1% wealthiest households, Scotland",
                                "Sample sizes, Scotland")),
                list(header = c("Absolute wealth gap between the wealthiest 10% and the least wealthy 40%",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Note: 2006-2008 estimates are based on the half of the sample that were asked the physical wealth questions.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Absolute (inflation-adjusted) wealth gap in £ between the wealthiest 10% and the least wealthy 40%, Scotland",
                                "Sample sizes, Scotland")),
                list(header = c("Gini coefficient by type of wealth",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Note: 2006-2008 estimates for total and physical wealth are based on the half of the sample that were asked the physical wealth questions.",
                                "Note: This analysis includes all households, whether or not they have each type of wealth.", 
                                "Source: Wealth and Assets Survey"),
                     titles = c("Gini coefficient of wealth inequality by type of wealth, Scotland",
                                "Sample sizes, Scotland")),

# * - individual characteristics -----------------------------------------------

                list(header = c("Individual wealth by age",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Individual wealth data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) individual wealth in £ by age, Scotland",
                                "Median age by age group, Scotland",
                                "Sample sizes by age, Scotland")),
                list(header = c("Individual wealth by economic status",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Economic status data has been available since 2014-2016.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) individual wealth in £ of adults by economic status, Scotland",
                                "Median age of adults by economic status, Scotland",
                                "Sample sizes by economic status, Scotland")),
                list(header = c("Individual wealth by qualification",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Individual wealth data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) individual wealth in £ of adults by qualification, Scotland",
                                "Median age of adults by qualification, Scotland",
                                "Sample sizes by qualification, Scotland")),
                list(header = c("Individual wealth by marital status",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Individual wealth data has been available since 2010-2012.",
                                "Note: The 'married' group also includes adults in Civil Partnerships; the 'divorced' group also includes separated adults and those in dissolved Civil Partnerships.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) individual wealth in £ of adults by marital status, Scotland",
                                "Median age of adults by marital status, Scotland",
                                "Sample sizes by marital status, Scotland")),
                list(header = c("Individual wealth by sexual orientation",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: The 'Other' group includes people who responded they were gay or lesbian, bisexual, or other.",
                                "Note: Sexual orientation data has been available since 2014-2016.",
                                "Note: Only adults who responded to the survey in person were included in the sexual orientation breakdown. No proxy-responses are included.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) individual wealth in £ of adults by sexual orientation, Scotland",
                                "Median age of adults by sexual orientation, Scotland",
                                "Sample sizes by sexual orientation, Scotland")),
                list(header = c("Individual wealth by religion",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Only adults who responded to the survey in person were included in the religion breakdown. No proxy-responses are included.",
                                "Note: Individual wealth data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) individual wealth in £ of adults by religion, Scotland",
                                "Median age of adults by religion, Scotland",
                                "Sample sizes by religion, Scotland")),

                list(header = c("Individual wealth by ethnicity",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: The 'minority ethnic' group combines all minority ethnic adults, including white ethnic minorities such as EU citizens and gypsy/travellers.", 
                                "Note: Only adults who responded to the survey in person were included in the ethnicity breakdown. No proxy-responses are included.", 
                                "Note: Individual wealth data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) individual wealth in £ of adults by ethnicity, Scotland",
                                "Median age of adults by ethnicity, Scotland",
                                "Sample sizes by ethnicity, Scotland")),
                list(header = c("Individual wealth by disability",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Individual wealth data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) individual wealth in £ of adults by whether they are disabled, Scotland",
                                "Median age of adults by whether they are disabled, Scotland",
                                "Sample sizes of adults by whether they are disabled, Scotland")),                
                list(header = c("Individual wealth by sex",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Individual wealth data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) individual wealth in £ of adults by sex, Scotland",
                                "Median age of adults by sex, Scotland",
                                "Sample sizes by sex, Scotland")),
                
# * - household characteristics ------------------------------------------------

                list(header = c("Household wealth by household income quintile",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Income data has been available since 2014-2016.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) household wealth in £ by equivalised household income (20%-band), Scotland",
                                "Median age of household head by household type, Scotland",
                                "Sample sizes by household type, Scotland")),
                list(header = c("Household wealth by household type",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: The 'Other' group contains households with additional adults such as non-dependent (grown-up) children.", 
                                "Note: 2006-2008 estimates are based on the half of the sample that were asked the physical wealth questions.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) household wealth in £ by household type, Scotland",
                                "Median age of household head by household type, Scotland",
                                "Sample sizes by household type, Scotland")),
                list(header = c("Household wealth by housing tenure",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Note: 2006-2008 estimates are based on the half of the sample that were asked the physical wealth questions.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) household wealth in £ by tenure, Scotland",
                                "Median age of household head by tenure, Scotland",
                                "Sample sizes by tenure, Scotland")),
                list(header = c("Household wealth in urban and rural areas",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: The data source is not representative of areas north of the Caledonian Canal and the Scottish islands; urban/rural breakdowns therefore require careful interpretation.",
                                "Note: 2006-2008 estimates are based on the half of the sample that were asked the physical wealth questions.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) household wealth in £ in urban and rural areas, Scotland",
                                "Median age of household head in urban and rural areas, Scotland",
                                "Sample sizes in urban and rural areas, Scotland")),
                
# * 3 Financial wealth & debt --------------------------------------------------
                
                list(header = c("Median financial wealth by financial wealth decile",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c(paste0("Median net financial wealth in £ by financial wealth decile, Scotland ", period),
                                paste0("Median age of household head by financial wealth decile, Scotland ", period),
                                paste0("Sample sizes by financial wealth decile, Scotland ", period))),
                list(header = c("Median financial wealth by financial wealth decile (time series)",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) net financial wealth in £ by financial wealth decile, Scotland",
                                "Median age of household head by financial wealth decile, Scotland",
                                "Sample sizes by financial wealth decile, Scotland")),
                list(header = c("Financial vulnerability and resilience",
                                "This worksheet contains four tables presented below each other, with the first table showing a quantitative measure of financial vulnerability, the second table providing sample sizes, the third showing a perceptive measure of financial vulnerability, and the final table again providing sample sizes.",
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Share of households whose savings, if replacing all income, would keep them above the poverty line for different lengths of time, Scotland",
                                "Sample sizes (households), Scotland",
                                "Share of adults who report they would be able to make ends meet for different lengths of time should they lose their main source of income, Scotland",
                                "Sample sizes (adults), Scotland")),

# *  - financial vulnerability -------------------------------------------------         

                list(header = c("Financial vulnerability by household type",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Households are considered financially vulnerable if their savings cover less than a month of income at the poverty line",
                                "Note: The 'Other' group contains households with additional adults such as non-dependent (grown-up) children.", 
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Proportion of households who are financially vulnerable by household type, Scotland",
                                "Composition of financially vulnerable households by household type, Scotland",
                                "Median age of household head by household type, Scotland",
                                "Sample sizes by household type, Scotland")),
                list(header = c("Financial vulnerability by economic status",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Households are considered financially vulnerable if their savings cover less than a month of income at the poverty line",
                                "Note: Economic status data has been available since 2014-2016.",
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Proportion of households who are financially vulnerable by economic status of the household head, Scotland",
                                "Composition of financially vulnerable households by economic status of the household head, Scotland",
                                "Median age of household head by economic status of the household head, Scotland",
                                "Sample sizes by economic status of the household head, Scotland")),
                list(header = c("Financial vulnerability by housing tenure",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Households are considered financially vulnerable if their savings cover less than a month of income at the poverty line",
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Proportion of households who are financially vulnerable by housing tenure, Scotland",
                                "Composition of financially vulnerable households by housing tenure, Scotland",
                                "Median age of household head by housing tenure, Scotland",
                                "Sample sizes by housing tenure, Scotland")),
                list(header = c("Financial vulnerability by income",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Households are considered financially vulnerable if their savings cover less than a month of income at the poverty line",
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Proportion of households who are financially vulnerable by equivalised household income quintile (20%-band), Scotland",
                                "Composition of financially vulnerable households by equivalised household income quintile (20%-band), Scotland",
                                "Median age of household head by equivalised household income quintile (20%-band), Scotland",
                                "Sample sizes by equivalised household income quintile (20%-band), Scotland")),
                list(header = c("Financial vulnerability by qualification",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Households are considered financially vulnerable if their savings cover less than a month of income at the poverty line",
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Proportion of households who are financially vulnerable by qualification of the household head, Scotland",
                                "Composition of financially vulnerable households by qualification of the household head, Scotland",
                                "Median age of household head by qualification of the household head, Scotland",
                                "Sample sizes by qualification of the household head, Scotland")),
                list(header = c("Financial vulnerability by age",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Households are considered financially vulnerable if their savings cover less than a month of income at the poverty line",
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Proportion of households who are financially vulnerable by age of the household head, Scotland",
                                "Composition of financially vulnerable households by age of the household head, Scotland",
                                "Median age of household head by age of the household head, Scotland",
                                "Sample sizes by age of the household head, Scotland")),
                list(header = c("Financial vulnerability by marital status",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Households are considered financially vulnerable if their savings cover less than a month of income at the poverty line",
                                "Note: The 'married' group also includes Civil Partnerships; the 'divorced' group also includes separated household heads and those in dissolved Civil Partnerships.",
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Proportion of households who are financially vulnerable by marital status of the household head, Scotland",
                                "Composition of financially vulnerable households by marital status of the household head, Scotland",
                                "Median age of household head by marital status of the household head, Scotland",
                                "Sample sizes by marital status of the household head, Scotland")),
                list(header = c("Financial vulnerability by whether there are children in the household",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Households are considered financially vulnerable if their savings cover less than a month of income at the poverty line",
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Proportion of households who are financially vulnerable by whether there are dependent children living in the household, Scotland",
                                "Composition of financially vulnerable households by whether there are dependent children living in the household, Scotland",
                                "Median age of household head by whether there are dependent children living in the household, Scotland",
                                "Sample sizes by whether there are dependent children living in the household, Scotland")),
                list(header = c("Financial vulnerability by disability",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Households are considered financially vulnerable if their savings cover less than a month of income at the poverty line",
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Proportion of households who are financially vulnerable by whether there are disabled household members, Scotland",
                                "Composition of financially vulnerable households by whether there are disabled household members, Scotland",
                                "Median age of household head by whether there are disabled household members, Scotland",
                                "Sample sizes by whether there are disabled household members, Scotland")),
                list(header = c("Financial vulnerability by sex",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Households are considered financially vulnerable if their savings cover less than a month of income at the poverty line",
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Proportion of households who are financially vulnerable by sex of the household head, Scotland",
                                "Composition of financially vulnerable households by sex of the household head, Scotland",
                                "Median age of household head by sex of the household head, Scotland",
                                "Sample sizes by sex of the household head, Scotland")),
                list(header = c("Financial vulnerability in urban and rural areas",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Households are considered financially vulnerable if their savings cover less than a month of income at the poverty line",
                                "Note: The data source is not representative of areas north of the Caledonian Canal and the Scottish islands;",
                                "urban/rural breakdowns therefore require careful interpretation.",
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Proportion of households who are financially vulnerable in urban and rural areas, Scotland",
                                "Composition of financially vulnerable households in urban and rural areas, Scotland",
                                "Median age of household head in urban and rural areas, Scotland",
                                "Sample sizes in urban and rural areas, Scotland")),
                list(header = c("Financial vulnerability by ethnicity",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Households are considered financially vulnerable if their savings cover less than a month of income at the poverty line",
                                "Note: The 'minority ethnic' group combines all minority ethnic households, including white ethnic minorities such as EU citizens and gypsy/travellers.", 
                                "Source: Wealth and Assets Survey and Family Resources Survey"),
                     titles = c("Proportion of households who are financially vulnerable by ethnicity of the household head, Scotland",
                                "Composition of financially vulnerable households by ethnicity of the household head, Scotland",
                                "Median age of household head by ethnicity of the household head, Scotland",
                                "Sample sizes by ethnicity of the household head, Scotland")),

# *  - unmanageable debt -------------------------------------------------------         

                list(header = c("Households with unmanageable debt",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Unmanageable debt data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt, Scotland",
                                "Median age of household head, Scotland",
                                "Sample sizes, Scotland")),
                list(header = c("Households with unmanageable debt by economic status",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Economic status data has been available since 2014-2016.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt by economic status of the household head, Scotland",
                                "Composition of households with unmanageable debt by economic status of the household head, Scotland",
                                "Median age of household head by economic status of the household head, Scotland",
                                "Sample sizes by economic status of the household head, Scotland")),
                list(header = c("Households with unmanageable debt by household type",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: The 'Other' group contains households with additional adults such as non-dependent (grown-up) children.", 
                                "Note: Unmanageable debt data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt by household type, Scotland",
                                "Composition of households with unmanageable debt by household type, Scotland",
                                "Median age of household head by household type, Scotland",
                                "Sample sizes by household type, Scotland")),
                list(header = c("Households with unmanageable debt by housing tenure",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Unmanageable debt data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt by housing tenure, Scotland",
                                "Composition of households with unmanageable debt by housing tenure, Scotland",
                                "Median age of household head by housing tenure, Scotland",
                                "Sample sizes by housing tenure, Scotland")),
                list(header = c("Households with unmanageable debt by ethnicity",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: The 'minority ethnic' group combines all minority ethnic households, including white ethnic minorities such as EU citizens and gypsy/travellers.", 
                                "Note: Unmanageable debt data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt by ethnicity of the household head, Scotland",
                                "Composition of households with unmanageable debt by ethnicity of the household head, Scotland",
                                "Median age of household head by ethnicity of the household head, Scotland",
                                "Sample sizes by ethnicity of the household head, Scotland")),                
                list(header = c("Households with unmanageable debt by age",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Unmanageable debt data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt by age of the household head, Scotland",
                                "Composition of households with unmanageable debt by age of the household head, Scotland",
                                "Median age of household head by age of the household head, Scotland",
                                "Sample sizes by age of the household head, Scotland")),
                list(header = c("Households with unmanageable debt by whether there are children in the household",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Unmanageable debt data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt by whether there are dependent children living in the household, Scotland",
                                "Composition of households with unmanageable debt by whether there are dependent children living in the household, Scotland",
                                "Median age of household head by whether there are dependent children living in the household, Scotland",
                                "Sample sizes by whether there are dependent children living in the household, Scotland")),
                list(header = c("Households with unmanageable debt by income",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Income data has been available since 2014-2016.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt by equivalised household income quintile, Scotland",
                                "Composition of households with unmanageable debt by equivalised household income quintile, Scotland",
                                "Median age of household head by equivalised household income quintile, Scotland",
                                "Sample sizes by equivalised household income quintile, Scotland")),                
                list(header = c("Households with unmanageable debt by marital status",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: The 'married' group also includes Civil Partnerships; the 'divorced' group also includes separated household heads and those in dissolved Civil Partnerships.",
                                "Note: Unmanageable debt data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt by marital status of the household head, Scotland",
                                "Composition of households with unmanageable debt by marital status of the household head, Scotland",
                                "Median age of household head by marital status of the household head, Scotland",
                                "Sample sizes by marital status of the household head, Scotland")),
                list(header = c("Households with unmanageable debt by disability",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Unmanageable debt data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt by whether there are disabled household members, Scotland",
                                "Composition of households with unmanageable debt by whether there are disabled household members, Scotland",
                                "Median age of household head by whether there are disabled household members, Scotland",
                                "Sample sizes by whether there are disabled household members, Scotland")),
                list(header = c("Households with unmanageable debt by qualification",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Unmanageable debt data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt by qualification of the household head, Scotland",
                                "Composition of households with unmanageable debt by qualification of the household head, Scotland",
                                "Median age of household head by qualification of the household head, Scotland",
                                "Sample sizes by qualification of the household head, Scotland")),                
                list(header = c("Households with unmanageable debt by sex",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Unmanageable debt data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt by sex of the household head, Scotland",
                                "Composition of households with unmanageable debt by sex of the household head, Scotland",
                                "Median age of household head by sex of the household head, Scotland",
                                "Sample sizes by sex of the household head, Scotland")),
                list(header = c("Households with unmanageable debt in urban and rural areas",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Unmanageable debt data has been available since 2010-2012.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with unmanageable debt in urban and rural areas, Scotland",
                                "Composition of households with unmanageable debt in urban and rural areas, Scotland",
                                "Median age of household head in urban and rural areas, Scotland",
                                "Sample sizes in urban and rural areas, Scotland")),
                
# * 4 Property wealth ----------------------------------------------------------
                
                list(header = c("Median property wealth by property wealth decile",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Includes households with property only.",
                                "Source: Wealth and Assets Survey"),
                     titles = c(paste0("Median net property wealth in £ by property wealth decile, Scotland ", period),
                                paste0("Median age of household head by property wealth decile, Scotland ", period),
                                paste0("Sample sizes by property wealth decile, Scotland ", period))),
                list(header = c("Median property wealth by property wealth decile (time series)",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Includes households with property only.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median (inflation-adjusted) net property wealth in £ by property wealth decile, Scotland",
                                "Median age of household head by property wealth decile, Scotland",
                                "Sample sizes by property wealth decile, Scotland")),
                
                list(header = c("Households with property wealth by household type",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: The 'Other' group contains households with additional adults such as non-dependent (grown-up) children.", 
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with property wealth by household type, Scotland",
                                "Composition of households with property wealth by household type, Scotland",
                                "Median age of household head by household type, Scotland",
                                "Sample sizes by household type, Scotland")),
                list(header = c("Households with property wealth by income",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Income data has been available since 2014-2016.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with property wealth by equivalised household income quintile, Scotland",
                                "Composition of households with property wealth by equivalised household income quintile, Scotland",
                                "Median age of household head by equivalised household income quintile, Scotland",
                                "Sample sizes by equivalised household income quintile, Scotland")),
                list(header = c("Households with property wealth by economic status",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: Economic status data has been available since 2014-2016.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with property wealth by economic status of the household head, Scotland",
                                "Composition of households with property wealth by economic status of the household head, Scotland",
                                "Median age of household head by economic status of the household head, Scotland",
                                "Sample sizes by economic status of the household head, Scotland")),
                list(header = c("Households with property wealth by qualification",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with property wealth by qualification of the household head, Scotland",
                                "Composition of households with property wealth by qualification of the household head, Scotland",
                                "Median age of household head by qualification of the household head, Scotland",
                                "Sample sizes by qualification of the household head, Scotland")),
                list(header = c("Households with property wealth by marital status",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: The 'married' group also includes Civil Partnerships; the 'divorced' group also includes separated household heads and those in dissolved Civil Partnerships.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with property wealth by marital status of the household head, Scotland",
                                "Composition of households with property wealth by marital status of the household head, Scotland",
                                "Median age of household head by marital status of the household head, Scotland",
                                "Sample sizes by marital status of the household head, Scotland")),
                list(header = c("Households with property wealth by age",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with property wealth by age of the household head, Scotland",
                                "Composition of households with property wealth by age of the household head, Scotland",
                                "Median age of household head by age of the household head, Scotland",
                                "Sample sizes by age of the household head, Scotland")),
                list(header = c("Households with property wealth by disability",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with property wealth by whether there are disabled household members, Scotland",
                                "Composition of households with property wealth by whether there are disabled household members, Scotland",
                                "Median age of household head by whether there are disabled household members, Scotland",
                                "Sample sizes by whether there are disabled household members, Scotland")),
                list(header = c("Households with property wealth by ethnicity",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: The 'minority ethnic' group combines all minority ethnic households, including white ethnic minorities such as EU citizens and gypsy/travellers.", 
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with property wealth by ethnicity of the household head, Scotland",
                                "Composition of households with property wealth by ethnicity of the household head, Scotland",
                                "Median age of household head by ethnicity of the household head, Scotland",
                                "Sample sizes by ethnicity of the household head, Scotland")),
                list(header = c("Households with property wealth by sex",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with property wealth by sex of the household head, Scotland",
                                "Composition of households with property wealth by sex of the household head, Scotland",
                                "Median age of household head by sex of the household head, Scotland",
                                "Sample sizes by sex of the household head, Scotland")),
                list(header = c("Households with property wealth in urban and rural areas",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Note: The data source is not representative of areas north of the Caledonian Canal and the Scottish islands; urban/rural breakdowns therefore require careful interpretation.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with property wealth in urban and rural areas, Scotland",
                                "Composition of households with property wealth in urban and rural areas, Scotland",
                                "Median age of household head in urban and rural areas, Scotland",
                                "Sample sizes in urban and rural areas, Scotland")),
                list(header = c("Households with property wealth by whether there are children in the household",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of households, the second showing the composition of households, the third showing age information, and the final table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of households with property wealth by whether there are dependent children living in the household, Scotland",
                                "Composition of households with property wealth by whether there are dependent children living in the household, Scotland",
                                "Median age of household head by whether there are dependent children living in the household, Scotland",
                                "Sample sizes by whether there are dependent children living in the household, Scotland")),
                
# * 5 Pension wealth -----------------------------------------------------------
                
                list(header = c("Median pension wealth by pension wealth decile",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Includes adults with private pensions only.",
                                "Source: Wealth and Assets Survey"),
                     titles = c(paste0("Median individual pension wealth in £ by pension wealth decile, Scotland ", period),
                                paste0("Median age by pension wealth decile, Scotland ", period),
                                paste0("Sample sizes by pension wealth decile, Scotland ", period))),
                list(header = c("Median pension wealth by pension wealth decile (time series)",
                                "This worksheet contains three tables presented below each other, with the first table showing the key estimates and the additional tables providing age information and sample sizes.",
                                "Note: Includes adults with private pensions only.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Median individual (inflation-adjusted) pension wealth in £ by pension wealth decile, Scotland",
                                "Median age by pension wealth decile, Scotland",
                                "Sample sizes by pension wealth decile, Scotland")),
                list(header = c("Adult men and women with private pensions (not yet in payment) by age",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates and the second table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c(paste0("Proportion of adult men and women with private pensions (not yet in payment) by age, Scotland ", period),
                                paste0("Sample sizes, Scotland ", period))),
                list(header = c("Adult men and women with private pensions (not yet in payment) (time series)",
                                "This worksheet contains three tables presented below each other, with the first table showing the proportion of adults, the second showing the composition of adults, and the final table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of adult men and women with private pensions (not yet in payment), Scotland",
                                "Composition of adults with private pensions (not yet in payment), Scotland",
                                "Sample sizes, Scotland")),
                list(header = c("Adult men and women with private pensions in payment by age",
                                "This worksheet contains two tables presented below each other, with the first table showing the key estimates",
                                "and the second table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c(paste0("Proportion of adult men and women with private pensions in payment by age, Scotland ", period),
                                paste0("Sample sizes, Scotland ", period))),
                list(header = c("Adult men and women with private pensions in payment (time series)",
                                "This worksheet contains three tables presented below each other, with the first table showing the proportion of adults, the second showing the composition of adults, and the final table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of adult men and women with private pensions in payment, Scotland",
                                "Composition of adults with private pensions in payment, Scotland",
                                "Sample sizes, Scotland")),
                
                list(header = c("Adults with private pension wealth by income",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of adults, the second showing the composition of adults, the third showing age information, and the final table providing sample sizes.",
                                "Note: Income data has been available since 2014-2016.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of adults with private pension wealth by equivalised household income quintile, Scotland",
                                "Composition of adults with private pension wealth by equivalised household income quintile, Scotland",
                                "Median age by equivalised household income quintile, Scotland",
                                "Sample sizes by equivalised household income quintile, Scotland")),
                list(header = c("Adults with private pension wealth by qualification",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of adults, the second showing the composition of adults, the third showing age information, and the final table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of adults with private pension wealth by qualification, Scotland",
                                "Composition of adults with private pension wealth by qualification, Scotland",
                                "Median age by qualification, Scotland",
                                "Sample sizes by qualification, Scotland")),
                list(header = c("Adults with private pension wealth by economic status",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of adults, the second showing the composition of adults, the third showing age information, and the final table providing sample sizes.",
                                "Note: Economic status data has been available since 2014-2016.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of adults with private pension wealth by economic status, Scotland",
                                "Composition of adults with private pension wealth by economic status, Scotland",
                                "Median age by economic status, Scotland",
                                "Sample sizes by economic status, Scotland")),
                list(header = c("Adults with private pension wealth by sexual orientation",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of adults, the second showing the composition of adults, the third showing age information, and the final table providing sample sizes.",
                                "Note: The 'Other' group includes people who responded they were gay or lesbian, bisexual, or other.",
                                "Note: Only adults who responded to the survey in person were included in the sexual orientation breakdown. No proxy-responses are included.",
                                "Note: Sexual orientation data has been available since 2014-2016.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of adults with private pension wealth by sexual orientation, Scotland",
                                "Composition of adults with private pension wealth by sexual orientation, Scotland",
                                "Median age by sexual orientation, Scotland",
                                "Sample sizes by sexual orientation, Scotland")),
                list(header = c("Adults with private pension wealth by marital status",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of adults, the second showing the composition of adults, the third showing age information, and the final table providing sample sizes.",
                                "Note: The 'married' group also includes adults in Civil Partnerships; the 'divorced' group also includes separated adults and those in dissolved Civil Partnerships.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of adults with private pension wealth by marital status, Scotland",
                                "Composition of adults with private pension wealth by marital status, Scotland",
                                "Median age by marital status, Scotland",
                                "Sample sizes by marital status, Scotland")),
                list(header = c("Adults with private pension wealth by age",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of adults, the second showing the composition of adults, the third showing age information, and the final table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of adults with private pension wealth by age, Scotland",
                                "Composition of adults with private pension wealth by age, Scotland",
                                "Median age by age, Scotland",
                                "Sample sizes by age, Scotland")),
                list(header = c("Adults with private pension wealth by religion",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of adults, the second showing the composition of adults, the third showing age information, and the final table providing sample sizes.",
                                "Note: Only adults who responded to the survey in person were included in the religion breakdown. No proxy-responses are included.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of adults with private pension wealth by religion, Scotland",
                                "Composition of adults with private pension wealth by religion, Scotland",
                                "Median age by religion, Scotland",
                                "Sample sizes by religion, Scotland")),
                list(header = c("Adults with private pension wealth by ethnicity",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of adults, the second showing the composition of adults, the third showing age information, and the final table providing sample sizes.",
                                "Note: The 'minority ethnic' group combines all minority ethnic adults, including white ethnic minorities such as EU citizens and gypsy/travellers.", 
                                "Note: Only adults who responded to the survey in person were included in the ethnicity breakdown. No proxy-responses are included.", 
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of adults with private pension wealth by ethnicity, Scotland",
                                "Composition of adults with private pension wealth by ethnicity, Scotland",
                                "Median age by ethnicity, Scotland",
                                "Sample sizes by ethnicity, Scotland")),
                list(header = c("Adults with private pension wealth by disability",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of adults, the second showing the composition of adults, the third showing age information, and the final table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of adults with private pension wealth by whether they are disabled, Scotland",
                                "Composition of adults with private pension wealth by whether they are disabled, Scotland",
                                "Median age by whether they are disabled, Scotland",
                                "Sample sizes by whether they are disabled, Scotland")),
                list(header = c("Adults with private pension wealth by sex",
                                "This worksheet contains four tables presented below each other, with the first table showing the proportion of adults, the second showing the composition of adults, the third showing age information, and the final table providing sample sizes.",
                                "Source: Wealth and Assets Survey"),
                     titles = c("Proportion of adults with private pension wealth by sex, Scotland",
                                "Composition of adults with private pension wealth by sex, Scotland",
                                "Median age by sex, Scotland",
                                "Sample sizes by sex, Scotland")))


# re-number headers and tables
names(headers) <- paste0("table_", seq(1, tables_n, 1))
names(all_tables) <- paste0("table_", seq(1, tables_n, 1))

# add prefixes to headings to mark them up as headings for post-processing
for (i in 1:tables_n) {
  headers[[i]]$header[1] <- paste0("H1-", headers[[i]]$header[1])
  headers[[i]]$titles <- paste0("H2-", headers[[i]]$titles)
}

# Run all worksheets -----------------------------------------------------------

# Open / create workbook
  if (file.exists(filename)) {wb <- loadWorkbook(filename)
  } else {wb <- createWorkbook()}

sapply(seq(1, tables_n, 1),
       function(x) {create_worksheet(sheet_no = x,
                                     wb = wb, 
                                     tables = all_tables[[x]], 
                                     headers = headers[[paste0("table_", x)]]$header, 
                                     titles = headers[[paste0("table_", x)]]$titles)})

# TOC --------------------------------------------------------------------------
createContentSheet(wb = wb,
                   title = "H1-Table of contents",
                   toptext = "Table numbers are the same as those used in the report at data.gov.scot/wealth.", 
                   headings = list(titles = c("H2-Total wealth", 
                                              "H2-Wealth inequality",
                                              "H2-Financial wealth & debt",
                                              "H2-Property wealth",
                                              "H2-Pension wealth"),
                                   location = c(0, 4, 26, 54, 67)))

# Readme sheet -----------------------------------------------------------------
createReadmeSheet(wb = wb,
                  notes = list("H1-Important notes" = c("Published: 18 February 2022",
                                                        "Next update: winter 2023/2024",
                                                        "The tables in this spreadsheet contain all estimates shown in the 'Wealth in Scotland' report. The report can be found here: data.gov.scot/wealth",
                                                        "Table numbers are the same as those used in the report. More detailed information on definitions and methodology can also be found in the report.",
                                                        "Estimates are based on Scotland data from the Wealth and Assets Survey. This survey is managed by the Office for National Statistics, who also published a report, available on the ONS website.",
                                                        "Each worksheet contains multiple tables, with the main table at the top, and additional tables below. The additional tables provide context such as sample sizes and average ages. Estimates based on smaller samples are less certain than those based on larger samples. Wealth estimates for older populations tend to be higher since wealth usually is built up during people's lifetime."),
                               "H2-Contact" = c("Maike Waldmann",
                                                "Scottish Government",
                                                "Communities Analysis Division",
                                                "Email: social-justice-analysis@gov.scot"),
                               "H2-Sample sizes" = c("Estimates are based on a small random sample of households in Scotland. Each worksheet includes information on the underlying sample size.",
                                                     "For estimated proportions or median amounts, the sample of the underlying base has to be at least 30 cases, for example 30 interviewed people, or 30 interviewed households (depending on the question). Where sample sizes are too small to produce a robust estimate, the estimate will be suppressed.", 
                                                     "Small differences in these estimates may not be meaningful and should be interpreted with care. This is because the estimates are based on a sample survey and could be slightly higher or lower if we interviewed a different sample of the population.", 
                                                     "In 2006-2008, only half of the households in the sample were asked about their physical wealth. Only these households are included in estimates for this period of total wealth and physical wealth.",
                                                     "In some breakdowns by individual or household characteristics, the 'All' category may not be based on the full Scottish sample, because we excluded a small number of individuals or households where the characteristic was unknown."),
                               "H2-Population estimates and coverage" = "Population estimates (the number of people or the number of households) are not available. However, we do publish average wealth, or proportions of households or individuals who have certain characteristics, for example unmanageable debt. Note also that the survey does not cover the regions north of the Caledonian Canal and the Scottish islands. It is therefore not representative of the whole of the Scottish population, and especially leaves out some of the most remote areas in Scotland.",
                               "H2-Rounding" = "Proportions are rounded to the nearest 1%. GBP amounts are rounded to the nearest £100.",
                               "H2-Unit of analysis and individual wealth" = "The unit of analysis is households or individuals, and is clearly stated in each table header. For individual wealth estimates, some types of wealth, such as pension wealth, can be clearly allocated to an individual. Other types of wealth, such as housing wealth, were split among household members. More detail is available in the report.",
                               "H2-Inflation adjustment" = "Wealth estimates have been deflated to April 2018 to March 2020 prices using the Office for Natinal Statistics' Consumer Prices Index including owner occupiers’ housing costs. More information is available in the report.",
                               "H2-Household head definition" = "The household head is the person with the highest income, or, for equal incomes, the older person."))


saveWorkbook(wb, filename, overwrite = TRUE)
